refactor(command): split command types into feature/command.rs
Move Command enum, types, parse helpers, and stream factories into
a dedicated module. feature/mod.rs now only contains the Feature trait.
